# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de création de sondages en temps réel avec Go et Gin Framework",
        "title": "QuickPoll API",
        "termsOfService": "https://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://quickpoll.example.com",
            "email": "support@quickpoll.example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/polls": {
            "post": {
                "description": "Create a new poll with options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Create a new poll",
                "parameters": [
                    {
                        "description": "Poll data",
                        "name": "poll",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/poll.CreatePollInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/poll.CreatePollOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/polls/{id}": {
            "get": {
                "description": "Get poll details with current results and vote counts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Get poll details",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Poll ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Poll details with results",
                        "schema": {
                            "$ref": "#/definitions/entity.Poll"
                        }
                    },
                    "400": {
                        "description": "Invalid poll ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Poll not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/polls/{id}/qr": {
            "get": {
                "description": "Generate QR code that links to the poll for easy sharing",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "polls"
                ],
                "summary": "Generate QR code for poll",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Poll ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR code image",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid poll ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to generate QR code",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/polls/{id}/vote": {
            "post": {
                "description": "Submit a vote for one or more options in a poll",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "votes"
                ],
                "summary": "Submit a vote",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Poll ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vote data with option IDs",
                        "name": "vote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.VoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vote submitted successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.VoteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or voting error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Option": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "order": {
                    "type": "integer",
                    "example": 0
                },
                "poll_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "text": {
                    "type": "string",
                    "example": "Go"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "vote_count": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "entity.Poll": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                },
                "created_by": {
                    "type": "string",
                    "example": "user123"
                },
                "description": {
                    "type": "string",
                    "example": "Choose your preferred programming language"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2024-01-16T10:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "multi_choice": {
                    "type": "boolean",
                    "example": false
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Option"
                    }
                },
                "require_auth": {
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "type": "string",
                    "example": "What's your favorite programming language?"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:00:00Z"
                }
            }
        },
        "handler.VoteRequest": {
            "type": "object",
            "required": [
                "option_ids"
            ],
            "properties": {
                "option_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "550e8400-e29b-41d4-a716-446655440001"
                    ]
                }
            }
        },
        "handler.VoteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "vote submitted successfully"
                }
            }
        },
        "poll.CreatePollInput": {
            "type": "object",
            "required": [
                "options",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Choose your preferred programming language"
                },
                "expires_in": {
                    "type": "integer",
                    "example": 60
                },
                "multi_choice": {
                    "type": "boolean",
                    "example": false
                },
                "options": {
                    "type": "array",
                    "maxItems": 10,
                    "minItems": 2,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Go",
                        "Python",
                        "JavaScript",
                        "Rust"
                    ]
                },
                "require_auth": {
                    "type": "boolean",
                    "example": false
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "What's your favorite programming language?"
                }
            }
        },
        "poll.CreatePollOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "qr_code_url": {
                    "type": "string",
                    "example": "http://localhost:8080/api/v1/polls/550e8400-e29b-41d4-a716-446655440000/qr"
                },
                "share_url": {
                    "type": "string",
                    "example": "http://localhost:8080/poll/550e8400-e29b-41d4-a716-446655440000"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}